requirements.txt
aiogram
psycopg2-binary
openai
python-dotenv
matplotlib
